Fix partial matching for filtering
Turns out that in order for the editor to match and filter the list based on a partial match for the resource identifier (the part after the `/`), it has to be the beginning of the `displayText`. Otherwise, the window is not filtered but bolded on matches.

By moving that part to the `displayTextPrefix` we get the desired filtering, and we switch the replacement behavior to use the `sortText` instead, which still contains the full string.