Improve versioning by using semver for major.minor.patch
We want to make it easy for the CI system to generate a VSIX that can be installed and upgraded more or less automatically.

In order for this to happen, the major.minor.patch version of the VSIX must be incremented with every build.

Also, we promote the `pr` metadata to a proper label, which makes the builds easier to tell apart from regular builds.