Fail build early and warn in IDE when resource not found
Instead of waiting for the full build to fail, warn users of missing resources much earlier, right in the editor, but also fail the build with the proper error, well before aapt runs.

In order for the build to also account for this new analyzer, we move the analyzer assembly to the MSBuild targets location, now under MSBuild\Xamarin\CodeAnalysis instead of the root Xamarin dir. Some tweaks were required on the VSIX project to properly collect the files and their target location.