Format analyzer message, properly detect category
For moving a string literal to a resource file, we should detect the property category the same way we do for the other analyzer, based on the `[Category]` attribute. This allows us to support more target files down the road, as well as detect `string` category on more places potencially (not just `Label` as we had before).

The message formats should include the value we'll be moving or we are reporting, to make it more friendly.