Produce a nuget package for the analyzer
This fixes an issue around additional files and compilation start
analyzers (see https://github.com/dotnet/roslyn/issues/35357). It
also makes it easier to test the functionality when on IDE-specific
extension is needed (unlike we have for custom completion provider
for Android).

That completion provider will need to be migrated to another IDE-only
analyzer assembly that is registered separately by the VSIX/Addin.